# go/snowplow/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This the total new Krateo backend.",
        "title": "SnowPlow API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/call": {
            "get": {
                "description": "Handle Resources",
                "produces": [
                    "application/json"
                ],
                "summary": "Call Endpoint",
                "operationId": "call",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource API Group and Version",
                        "name": "apiVersion",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource Plural",
                        "name": "resource",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pagination desired page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination desired per page items",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "JSON encoded map of extra params",
                        "name": "extras",
                        "in": "query"
                    },
                    {
                        "description": "Object",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    }
                }
            },
            "put": {
                "description": "Handle Resources",
                "produces": [
                    "application/json"
                ],
                "summary": "Call Endpoint",
                "operationId": "call",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource API Group and Version",
                        "name": "apiVersion",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource Plural",
                        "name": "resource",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pagination desired page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination desired per page items",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "JSON encoded map of extra params",
                        "name": "extras",
                        "in": "query"
                    },
                    {
                        "description": "Object",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    }
                }
            },
            "post": {
                "description": "Handle Resources",
                "produces": [
                    "application/json"
                ],
                "summary": "Call Endpoint",
                "operationId": "call",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource API Group and Version",
                        "name": "apiVersion",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource Plural",
                        "name": "resource",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pagination desired page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination desired per page items",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "JSON encoded map of extra params",
                        "name": "extras",
                        "in": "query"
                    },
                    {
                        "description": "Object",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    }
                }
            },
            "delete": {
                "description": "Handle Resources",
                "produces": [
                    "application/json"
                ],
                "summary": "Call Endpoint",
                "operationId": "call",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource API Group and Version",
                        "name": "apiVersion",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource Plural",
                        "name": "resource",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pagination desired page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination desired per page items",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "JSON encoded map of extra params",
                        "name": "extras",
                        "in": "query"
                    },
                    {
                        "description": "Object",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    }
                }
            },
            "patch": {
                "description": "Handle Resources",
                "produces": [
                    "application/json"
                ],
                "summary": "Call Endpoint",
                "operationId": "call",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource API Group and Version",
                        "name": "apiVersion",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource Plural",
                        "name": "resource",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource namespace",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pagination desired page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination desired per page items",
                        "name": "perPage",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "JSON encoded map of extra params",
                        "name": "extras",
                        "in": "query"
                    },
                    {
                        "description": "Object",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Status"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.Status": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "code": {
                    "description": "Suggested HTTP return code for this status, 0 if not set.",
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "message": {
                    "description": "A human-readable description of the status of this operation.",
                    "type": "string"
                },
                "reason": {
                    "description": "A machine-readable description of why this operation is in the\n\"Failure\" status. If this value is empty there\nis no information available. A Reason clarifies an HTTP status\ncode but does not override it.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/response.StatusReason"
                        }
                    ]
                },
                "status": {
                    "description": "Status of the operation.\nOne of: \"Success\" or \"Failure\".",
                    "type": "string"
                }
            }
        },
        "response.StatusReason": {
            "type": "string",
            "enum": [
                "",
                "Unauthorized",
                "Forbidden",
                "NotFound",
                "Conflict",
                "Gone",
                "Invalid",
                "Timeout",
                "TooManyRequests",
                "BadRequest",
                "MethodNotAllowed",
                "NotAcceptable",
                "RequestEntityTooLarge",
                "UnsupportedMediaType",
                "UnprocessableEntity",
                "InternalError",
                "ServiceUnavailable"
            ],
            "x-enum-varnames": [
                "StatusReasonUnknown",
                "StatusReasonUnauthorized",
                "StatusReasonForbidden",
                "StatusReasonNotFound",
                "StatusReasonConflict",
                "StatusReasonGone",
                "StatusReasonInvalid",
                "StatusReasonTimeout",
                "StatusReasonTooManyRequests",
                "StatusReasonBadRequest",
                "StatusReasonMethodNotAllowed",
                "StatusReasonNotAcceptable",
                "StatusReasonRequestEntityTooLarge",
                "StatusReasonUnsupportedMediaType",
                "StatusUnprocessableEntity",
                "StatusReasonInternalError",
                "StatusReasonServiceUnavailable"
            ]
        }
    }
}